requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
comm==0.1.4
cryptography==41.0.5
debugpy==1.8.0
decorator==5.1.1
ecdsa==0.18.0
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.0
greenlet==3.0.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.17.1
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.5.0
jupyter_core==5.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
platformdirs==3.11.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.16.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pyzmq==25.1.1
requests==2.31.0
rich==13.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
tomli==2.0.1
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
wcwidth==0.2.9
